Lint, profile type, profilePage, editProfile
