+++ src/UniSwap/Components/Header.css
@@ -28,7 +28,6 @@
     margin: 0px;
     font-size: 4em;
 }
-
 .home-icon-image {
     width: 40px;
     height: 40px;
