+++ src/UniSwap/Components/Header.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import "./Header.css";
 import marketPlace from "../Images/marketplace_icon.png";
 import mag_glass from "../Images/magnifying_glass_icon.png";
@@ -6,6 +6,19 @@ import { Link } from "react-router-dom";
 
 export default function Header(
 ) {
+  const [width, setWidth] = useState(window.innerWidth);
+  useEffect(() => {
+    const handleResize = () => {
+      setWidth(window.innerWidth);
+    };
+  
+    window.addEventListener('resize', handleResize);
+  
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
   const HomeIcon = (
     <Link to="/home">
       <div 
@@ -42,11 +55,11 @@ export default function Header(
   )
 
   return(
-    <div className="row" style={{ alignItems: 'center' }}>
+    <div className="row" style={{ alignItems: 'center', flexWrap:'nowrap', height: 96 }}>
       {HomeIcon}
       <AccountButton />
       {SearchBar}
-      <h3 className="adjustedFont uniswapLogoMini titleColor">UniSwap</h3>
+      {width > 1150 && <h3 className="adjustedFont uniswapLogoMini titleColor">UniSwap</h3>}
     </div>
   )
 }
