+++ src/UniSwap/Profile/index.css
@@ -18,7 +18,6 @@
     border: 2px solid #B99D61;
     margin-right: 10px;
     margin-left: 10px;
-
 }
 .profile {
     margin-top: 0px;
@@ -41,8 +40,7 @@
     position: relative;
     display: flex;
     padding: 20px;
-    width: 1196px;
-    height: 650px;
+    margin-left: 90px;
 }
 
 .headerContainer {
@@ -63,13 +61,12 @@
     margin-top: 20px;
 }
 
-.transparent-textarea {
+.profileBio {
     position: relative;
+    text-align: center;
     padding: 10px;
-    border: 1px solid #000;
-    border-radius: 5px;
-    background-color: rgba(255, 255, 255, 0.5);
-    color: #000;
+    color: #6F6F6F;
+    font-size: 1.1rem;
     width: 516px;
     height: 260px;
 }
@@ -121,5 +118,8 @@
 
 .edit-button {
     color: #000;
-    font-size: 20px;
+    background-color: transparent;
+    width: 25px;
+    align-self: center;
+    border-radius: 0px;
 }
